backend/requirements.txt
fastapi
uvicorn
python-dotenv
crewai
crewai_tools
langchain
langchain-openai
langchain-anthropic
langchain-google-vertexai
requests
websockets
weaviate-client
neo4j
sentence-transformers
pandas
lark
psycopg2-binary
sqlalchemy
